Added actions.py containg action methods
Currently methods are a test debugging message and a method to
send me the IP address of the PI that this script is running on.
scanner.py uses getattr to convert the string from the mapped UID to
a function that is callable by python.  Note that the function name should
match the string in the .csv file.